requirements.txt
tinytag
pygame==1.9.6
mutagen
Pillow
tkinter
webbrowser